stdout redirection now properly logged
Was forced to move RedirectStream to his own file to prevent a circular dependency
